[dashboard] remove start-with-options feature flag
